Gaelo2

+++ GaelO2/GaelO2/app/GaelO/UseCases/DeleteUser/DeleteUser.php
@@ -3,11 +3,15 @@
 namespace App\GaelO\UseCases\DeleteUser;
 
 use App\GaelO\Constants\Constants;
+use App\GaelO\Exceptions\GaelOException;
+use App\GaelO\Exceptions\GaelOForbiddenException;
+use App\GaelO\Exceptions\GaelONotFoundException;
 use App\GaelO\Interfaces\PersistenceInterface;
 use App\GaelO\Services\AuthorizationService;
 use App\GaelO\Services\TrackerService;
 use App\GaelO\UseCases\DeleteUser\DeleteUserRequest;
 use App\GaelO\UseCases\DeleteUser\DeleteUserResponse;
+use Exception;
 
 class DeleteUser {
 
@@ -19,20 +23,43 @@ class DeleteUser {
 
     public function execute(DeleteUserRequest $deleteRequest, DeleteUserResponse $deleteResponse) : void {
 
-        $this->authorizationService->isAdmin($deleteRequest->currentUserId);
+        try{
 
-        $this->persistenceInterface->delete($deleteRequest->id);
-        $deleteResponse->status = 200;
-        $deleteResponse->statusText = 'OK';
+            $this->checkAuthorization($deleteRequest->currentUserId);
 
-        $actionsDetails = [
-            'deactivated_user'=>$deleteRequest->id
-        ];
+            if ( ! $this->persistenceInterface->isExistingId($deleteRequest->id)){
+                throw new GaelONotFoundException('Non Exiting User Id');
+            };
 
-        $this->trackerService->writeAction($deleteRequest->currentUserId,
-                                Constants::TRACKER_ROLE_USER, null, null,
-                                Constants::TRACKER_EDIT_USER, $actionsDetails);
+            $this->persistenceInterface->delete($deleteRequest->id);
 
+            $actionsDetails = [
+                'deactivated_user'=>$deleteRequest->id
+            ];
+
+            $this->trackerService->writeAction($deleteRequest->currentUserId,
+                                    Constants::TRACKER_ROLE_USER, null, null,
+                                    Constants::TRACKER_EDIT_USER, $actionsDetails);
+
+            $deleteResponse->status = 200;
+            $deleteResponse->statusText = 'OK';
+
+        }catch ( GaelOException $e ) {
+            $deleteResponse->status = $e->statusCode;
+            $deleteResponse->statusText = $e->statusText;
+            $deleteResponse->body = $e->getErrorBody();
+
+        } catch (Exception $e){
+            throw $e;
+        };
+
+    }
+
+    private function checkAuthorization(int $userId) : void {
+        $this->authorizationService->setCurrentUserAndRole($userId);
+        if( ! $this->authorizationService->isAdmin()) {
+            throw new GaelOForbiddenException();
+        };
     }
 
 }
